data types fixing

+++ internal/globals/globals.go
@@ -10,7 +10,7 @@ type ConfigStruct struct {
 	HttpPort uint   `env:"AS_HTTP_PORT" envDefault:"80"`
 	Url      string `env:"AS_HTTP_URL" envDefault:"http://localhost:80"`
 	BotToken string `env:"AS_TELEGRAM_BOT_TOKEN" envDefault:""`
-	ChatID   int    `env:"AS_TELEGRAM_CHAT_ID" envDefault:"0"`
+	ChatID   int64  `env:"AS_TELEGRAM_CHAT_ID" envDefault:"0"`
 }
 
 // GLOBAL VARIABLES
